fdmmhydra/requirements.txt
flask==2.3.1
flask_markdown
flask_admin
flask_sqlalchemy
flask_wtf
flask.ext
flask_security
pytz
flask_apscheduler
packaging
requests
email_validator
sqlalchemy==1.4.45
psutil
pandas
folium
geopy
seaborn
pyyaml